Bumping version to 4.9.41.rc.004. Previous Commit message: "fix message admin helpers for acceptance tests [ci skip]
I recently implemented a change for the `messagesActivityListTableCellSelectorFor` method that would only return the designated numbered match from what matched the locator lookup.  However, I forgot that there are times where we DO want to return all matches (check for count of items etc) so I tweaked the logic to allow for that.
"