don’t round running total [touch: 10817]
This was cruft from an initial fix I had tried in the ticket for the original issue. I missed removing this once I realized the original issue was fixed by simply ensuring the accounting.js library was initialized with the correct config.

So what ended up happening is that I was sending into `accounting.formatNumber` a precision set number in the native js format (i.e. `1.00`) which accounting js then saw as 100 becuase the config had a period as a thousands separator.