show notices on page load [touch: 10838]
espresso_core.js has some code that executes before ajax requests that hides any existing notices on the page.  This means that no success or error notices ever get shown on the country settings page because it executes an ajax request right away on page load.

The fix is to check to see if there is a notice on the page and if there is then show it.

The fix also requires removing the style attribute because it has the style that is partway through the fadeout animation also done via expresso_core.js