If caps are added or removed, flush roles cached on WP_User [touch: 10803]
The current user loaded by WordPress caches internally in its object any roles that user has which in turn results in the caps from that role at the time of instantiation being registered for that user.  However if the assigned to a role change later in the request then that current user object won’t have those new caps updated until its instantiated again or a new request.

So the fix is that whenever EE_Capabilities adds or removes caps to a role, we ensure that any roles in the current user logged in have their caps reloaded and registered so changes are applied immediately.