implement new Context entity when calling EE_Registration::set_status [touch: 10992]
I thought the best thing to do here would be to utilize whatever the current page/route is for the context slug.  That way no matter where `_set_registration_status` is called from, it should set something useful as the Context.

The downside of course is that if page_slug or `_req_action` ever change in the future, then anything relying on those contexts will not work.  However, page_slug and _req_action are pretty critical identifiers used throughout the admin so the likelihood of us changing this in any future refactor (and not preserving some sort of back compat) is close to nil.