preserver original contexts and reset addressees after generating [touch: 11019]
The `_contexts` property needs to be reset after _contexts is overridden and used for generating the addressee objects so any future calls to the message type will have the full original contexts array available.

The `_addressees` property is only used for accumulating EE_Addressee objects for all contexts when processing the data.  Thus when that has been done, it should be reset otherwise it will continue to accumulate addressee data for further calls to this message type in the same request.