verify payment object has an ID [touch: 8880]
I think the incoming data handler was creating a default empty payment object as a way of preventing errors elsewhere.  So we have to check first whether the incoming payment object has an ID().  If it doesn’t then we know that we need to get the last payment from the transaction.