restore skipped test [touch: 11368]
This test was skipped because of a work in progress ticket that was intended to fix it.  However I also tweaked how this test worked. `EEH_DTT_Helper::get_timezone_string_from_gmt_offset` is intended to return a string that can be passed into a DateTimeZone object constructor NOT a string that can be passed as the value for the `timezone_string` WordPress option.  So, I modified the tests to instead update `gmt_offset` option in WP for the purpose of testing.  This more closely mirrors real world activity as sites using UTC offsets will have the `gmt_offset` option set NOT `timezone_string`.
I also updated the phpdocs for EEH_DTT_Helper::get_timezone_string_from_gmt_offset so its purpose is clearer.