remove usage of icon [touch: 11333]
WordPress core strips any html tags when the label is generated from the same string for the screen options that toggle whether the column displays or not.  This means that using icons can’t be done (easily) for columns.  Hence the necessity to switch back to a text label.

The choice of text was okayed by Seth in slack.