fix bad test [touch: 11469]
Make sure that timezone set on the models is the same as the timezone used to create the dates/times for the test.  Otherwise this test will fail every year around the end of march when the timezone difference between `America/Vancouver` and `UTC` causes last months date to be output as March 1 not February 28 because of the UTC difference.