Bumping version to 4.9.62.rc.027. Previous Commit message: "Merge pull request #359 from eventespresso/BUG/fix-off-by-one-second-phpunit-tests
Improve tests to allow passes when times are within one minute (fixes #341)

Precision is tricky in unit tests because things can affect the second precision when tests are run all in one go. So changes for test artifacts to be off by one second are greater. I think it’s sufficient to have a time tolerance of one minute difference within the unit tests for any comparisons between date and time strings."