Bug/fix error in console with jed initialization (fixes #372) (#393)
* Wrap import of @wordpress/i18n with eejs.i18n and expose as @eventespresso/i18n

This is needed because currently wp.i18n is not a part of wp core.  By wrapping the official pacakge, we should preserve all the benefits of wp.i18n and allow `i18n` to “just work” with our js via the exposed `eejs` global (which is necessary for initializing jed with the built strings for the loaded js).