Bumping version to 4.9.63.rc.011. Previous Commit message: "Switch webpack-assets-manifest plugin config to use the assets configuration vs merge (#486)
We were experiencing some odd behaviour in upstream branches where the number of compilation configurations in webpack seems to be impacting the final `build-manifest.json` output.  It appears it was related to using `merge` in this plugins configuration instead of the `assets` option (which is explicitly FOR multiple compilation configurations).  Since this seemed to fix the issue we’re rolling with it."