Add Value Objects for DateTime and Duration (#668)
* install moment and moment-timezone

* expose default timezone data from server

* expose moment-timezone on the moment object instead of vanilla moment

* expose TIMEZONE_CONFIG on eejs

* add custom exceptions for date-time related things

* various assertion/validation methods for date-time things

* add Duration vo and tests

* initial pass at DateTime vo

* add better validation of ISO8601 strings

Also accounts for the difference between a Duration string and a Datetime string.

* better account for incoming data type on construct

Also some other minor tweaks

* add additional tests

* fix typo in jsdoc

* Add jsdocs and a few tweaks to code for DateTime vo

* Add additional validation to assertion methods

* switch `moment` import to `momenttz` import

- this includes what is exposed on the vendor file which is now `eejs.vendor.momenttz` instead of `eejs.vendor.moment`.

This is required because of conflics with other libraries that might be importing `moment` from the `moment` library and not the `moment-timezone` library.

* improve duration and update its test files

* improve and add additional assertions

* expose and export new errors correctly

- also updates the eej.mock to expose the new errors.

* update DateTime vo and add tests

* rename datetime.js vo main file

- makes imports/exports a bit easier.

* export DateTime and Duration vos

* remove unnecessary import

* fix conditional check

* fix linting issues

* import moment-timezone not moment

This expands on earlier work with this import switch and reverts the custom `momenttz` module name.  This helps keep things clearer.  The Webpack config still handles mapping `momenttz` (for `eejs.vendor` global`) the the appropriate moment bundle.

* expose locale data

* set appropriate defaults for timezoneConfig (from server)

* add a defaults file for date-time vo

* implement DEFAULT_LOCALE

* implement Defaults and update VO class name

* add `ServerDateTime` vo

This allows for instantiating DateTime Vo’s that automatically assume what may be set as timezone string or offset on the server.

* improve eejs mock

* update dev build files

* change method used

travis is failing with `toThrow` so maybe `toThrowError` will work better?

* tweak travis config for node version

We’ll see if this fixes the odd issue on travis where custom error tests are failing.

* change name of property

* try adding name property see if that fixes travis fails

* remove php version set on license_check stage

* try some other fix attemps for travis build fails

* update travis yml again and more tweaks to exceptions

… so, I discovered the travis.yml had `nvm install && nvm use` so even though the build started with latest node, it was reverting to the LTS version.

* remove unnecessary npm install commands

* use ES5 style for exception

* fix linting issues

* fix linting issues

* use es5 for InvalidArgument

babel does not transpile classes extending native js classes (Error, Date etc) which results in errors in the travis environment (still not sure why it wasn’t happening locally).  Switching InvalidArgument to use prototype for extending Error should fix that and `InvalidArgument` can then be extended via ES6 and babel will transpile that.

* use DateTime format from site as the default `toFormat` argument.

* fix validation/assertion for locale

* include argument with `toISO` for returning the ISO string in UTC (default)

- also adds test coverage.

* fix jsdoc

* add datetime vo documentation

* use default locale (from site) for Duration

* fix js docs

* tweak datetime vo documentation

* Add Duration value object documentation

* add `DEFAULT_VALID_LOCALE`

This prevents errors thrown when the server locale doesn’t have an equivalent in moment.  While moment has a large number of locales, there is the possibility that a i18n locale being used on WordPress might not available in moment so this accounts for that and fallsback to `en`

* make sure ServerDateTime always returns an instance of itself rather than DateTime

* add documentation for ServerDateTime

* fix linting issues

* don’t export momenttz for eejs.vendor.

Event though we’re importing moment-timezone, it is essentially a decorator over the moment object, so there’s no need to differentiate it with the name.  This also preserves back compat on the off chance someone is utilizing `eejs.vendor.moment` already.

* update dev build

* fix jsdoc

* add missing jsdocs

* add missing quotes in example

* export ServerDateTime as well

* improve message construction for custom errors

* fix jsdocs

* include additional assertions and validation methods for duration

* fix typo

* ensure `fromISO*`  methods do not accept empty strings.

There are alternatives for construction a `DateTime` equivalent to “now”:

1. Use the constructor with an empty string
2. Use `fromJSDate`
3. Use `fromLocal` and then set timezone or offset if needed to change.

* capitalize UTC in jsdocs

* expand error text

* fix jsdoc

* fix error text

* restructure logic to remove unnecessary conditional

* fix jsdoc

* fix jsdocs

* restructure conditional for readability

* fix typo in symbol description

* fix typo and spacing

* validate incoming simple object

* fix documentation

* remove allowing a moment object to be passed into the constructor

and all the related fixes/updates as a result.

* update documentation

* temporary debug for travis

* fix test and logic

so if you pass something like ‘2018-12-25T10:15:00.200Z’ to `moment()` (which is the output of `(new Date()).toISOString()` moment does not treat that as equivalent to +00:00 but instead considers it being in local time.  This was exposed on travis.

* rename to fromMilliseconds

* add link to wikipedia entry

* modify method names so its more clear they are iso related

* improve the Duration.isValid() check.

* modify js doc