don’t use factory in eventespresso/core reducers/actions
This creates too much of a dependency.  Instead lets EXPECT entity instances for core and client code is responsible for creating them.

In testing I realized this is necessary for the way the resolver generators work if I want to make `eventespresso/core` the authority for all the other linked stores.