AutoFormRow
This component determines whether the incoming form element is an input, label, or row component and responds accordingly.
FormRow components from a form layout strategy will simply be returned as is, since they require no further manipulation.
Inputs and Labels will be passed to the supplied form layout strategy AutoColumnRow component for further processing.
If a Label component is not provided but the supplied Input has a "label" prop, then that will be used to generate an InputLabel using the supplied form layout strategy component.