Bumping version to 4.9.70.rc.003. Previous Commit message: "introduce default date coercing (#743)
* introduce default date coercing

Creating a DateTime (or DbSafeDateTime) object with `0000-00-00 00:00:00` results in a internal timestamp of `-0001-11-30 00:00:00.000000`.  This results in potential downstream problems when we go to output this because even when you output back to mysql date format you’ll get `-0001-11-30 00:00:00` which is DIFFERENT than what was sent in.

To resolve this, if we coerce those date and time strings to `0000-01-01 00:00:00` it will result in an internal timestamp of `0000-01-01 00:00:00.000000` which is better because it represents a more legible value as well as is consistent with what the object was created with.

Downstream this is more friendly when dates are converted to ISO strings to pass along on REST api packages etc.

* add typehinting for DateTimeZone

* remove typehinting, travis is clear it can’t be used.

* implement method in prepare_for_set as well

* ensure new method is returning instance of DbSafeDateTime

* fix expectation on tests

* preserve behaviour for constructing DateTime::createFromFormat
"