Increase default date
The problem with using `0000-01-01 is that if a DateTime is created in UTC and then the timezone is set to something that has a negative offset, the date will get set back before `0000-00-00` and thus end up as something like `-001-12-31`.  To prevent any possibility of this happening its better to set the default to the third day.