Improve PhpStorm code assistance on aliased javascript imports (#768)
PHPStorm doesn't provide code assistance for any imports we use in our js modules that are configured in webpack to point to externals.  For example in webpack, `@eventespresso/components` points to `eejs.components`.  This means that whenever we use something like this in our modules, PhpStorm doesn't know how to resolve it and thus we don't get things like code completion etc for that import:

```
import { isSchema } from '@eventespresso/validators';
```

This pull fixes that by exposing a special `phpstorm-webpack.config.js` file that is not used in any builds, but configured specifically to help PHPStorm with resolving aliased imports.

Included documentation in the commit provides instructions on how to implement.