Up the requirements in the readme.txt (#797)
On WordPress.org I noticed this today in the listing for our decaf plugin:

![screenshot](https://www.evernote.com/l/AAONAdY3bSZH-ITbDsjKcc8E68tLRBJMcXEB/image.png)

However, our [requirements page](https://eventespresso.com/requirements/) lists EE4 as requiring WordPress 4.5.  This pull thus updates the readme.txt to make it consistent with what we publish on the main website.

I've submitted this as a pull on the off chance decaf has a different requirement level than caf that I was not aware of.  If the pull is okay it can just be merged right into master.