fix eventespress/core state to account for plain object root
With the use of  Redux.combineReducers  (via wp.element), the root state is actually a plain object with all the individual state slices implementing immutable.js.  In the tests I wrote, I was considering the entire state being an instance of Immutable.Map when that’s not actually the case!  So the tests passed for that assumption but not really testing real-world use-case.

Although we _could_ implement our own (or a redux-immutable) combineReduceers to make the root state an Immutable.Map()… that could have repercussions down the road for other bolt-on things exposed by WP for the data stores that expect the state root to be a plain obejct.  I figured it’s safer and more future proof to simply have the state slices that matter using immutable (as the reducers/selectors are what will trigger re-renders).