remove dirty relations index actions and reducers
The index is still in the state, but exposing this level of granularity to clients created potential problems with order of execution etc.   This state tree is more of an optimization (lookup) trick for internal use so I’ve taken care of automatically changing its state internally via the reducer rather than controlled via its own action.