Fix current date and time always showing in date selector (Event Attendees block) (#911)
The Event Attendees block is always showing the current datetime.  I think this happened as a result of the model-entity work being merged in.  The fix involved correctly accounting for the new DateTime model-entities which include usage of `DateTime` value objects for the date values.

Tests are updated as well.