Bumping version to 4.9.78.rc.020. Previous Commit message: "More type proofing for pue client (fixes #803) (#933)
## Problem this Pull Request solves

See #803 for details.  There's already some type proofing in the places the warnings are being reported for but if there was bad data already in the database it could account for what that user was reporting.  Thus, the commits in this pull should handle even the bad data pulled from the db."