Bumping version to 4.9.78.rc.027. Previous Commit message: "Optimize usage of react key for AttendeeListItem component (#951)
## Problem this Pull Request solves
<!-- Please describe your changes in the context of the problem they solve -->

Keys in a react list element should be something that identifies that element as unique across re-renders.  This means in the case of list elements based off of entities its best to use the ID of that entity because it ensures the list item is not re-rendered if the id doesn't change.  In the case where an array of attendees might have an item that does not have an id, `cuid` is used instead.

"