Bumping version to 4.9.79.rc.007. Previous Commit message: "Add checkin action handling to `eventespresso/core` store (#965)
## Problem this Pull Request solves
<!-- Please describe your changes in the context of the problem they solve -->

While working on https://github.com/eventespresso/eea-barcode-scanner/pull/7 I realized that there is need for handling the toggling of checkins.  Checkins are a unique data model in that whenever a checkin is toggled for a given registration and datetime, a new record is created.  This allows for an audit trail of when a registration checked in and out of an event.

For the purpose of the barcode scanner, the ui presents a button for both indicating the current toggle action available (requiring the latest checkin record) and initiating toggling to the relevant checkin state (i.e. checked in or out).

The complexity of the check-in model is represented by the creation of an [rpc endpoint](https://github.com/eventespresso/event-espresso-core/blob/master/docs/C--REST-API/ee4-rest-api-rpc-checkin.md) specifically for handling checkins.

While the current selectors and actions available in our stores _could_ be used for handling things (with the exception of the toggle checkin endpoint), consuming code would have to do more work to correctly handle checkin changes since when a checkin is "toggled" via the rpc endpoint, a NEW record is returned (not an update to an existing record) - I'll add more details at another time just how much more code this creates for consumer.

For the purpose of simplifying things this pull seeks to:

- Add a new selector (and resolver) to `eventespresso/core` that is used for retrieving the latest checkin record `getLatestCheckin( registrationId, datetimeId );`
- Add a new action (generator) to `eventespresso/core` that is used for `receiving` the latest checkin: `receiveLatestCheckin( checkinEntity )`.
- Add a new action (generator) to `eventespresso/core` that is used for toggling a checkin (and in turn updates necessary store state via further dispatches): `toggleCheckin( checkinEntity )`

Under the hood these will use both new code, and existing actions for recording the checkin record in the state.

The shape of the new state tree (using Immutable!) will be something like this:
```
{
	modelSpecific: {
		getLatestCheckin: {
			[JSON.stringify([ registration.id, datetime.id] )]: checkInEntity.id,
		}
	}
}
```

So to be clear it simply stores the checkin id and the selector would retrieve the actual entity from the `entities` tree in the state.

These serve the goal of simplifying the api for consumers subscribing to the latest checkin because they would simply have to do something like this.

```js
export default compose( [
	withSelect( ( select, ownProps ) => {
		const { registration, datetime } = ownProps;
		return {
			checkin: select( 'eventespresso/core' ).getLatestCheckin(
				registration.id,
				datetime.id
			),
		}
	} ),
	withDispatch( ( dispatch ) => {
		return {
			onChange: ( checkin ) => dispatch( 'eventespresso/core' )
				.toggleCheckin( checkin.regId, checkin.dttId ),
		}
	} ),
] )( WrappedComponent );
```

This pull introduces the following:

- new "model specific" interface for adding new specific type model queries.  This includes a new action (`receiveSelectorValue`), a new selector (`getSelectorValue`), and new reducer.  These are interfaces to the model-specific state that can be used by more concrete model specific queries.
- The first model specific item to use this new interface is the interaction with latest checkin.  This includes new actions (`receiveLatestCheckin`, `toggleCheckin`), a new selector and resolver (`getLatestCheckin`).  Reducer is unneeded because that's handled via the model-specific interface.
- added a new control to base-controls for handling selects that have a resolver.  This is a better implementation than calling a generator from `base-resolvers`.  After this pull is merged, I'll do another pull to implement this wherever else we are using `base-resolvers` (which impacts more existing things so will need more testing).
- added unit test coverage.
- added documentation.

## How has this been tested
<!-- Please describe in detail how you tested your changes and how testing can be reproduced -->
<!-- Include details of your testing environment, and tests ran to see how your changes affect other areas of code -->
<!-- Include any notes about automated tests you've written for this pull request.  Pull requests with automated tests are preferred. -->

* [x] Added new unit tests, ensure existing tests don't fail.
* [x] Monkey testing in the browser to ensure new selects/actions work as expected.

## Checklist

* [x] I have read the documentation relating to systems affected by this pull request, see https://github.com/eventespresso/event-espresso-core/tree/master/docs
* [x] User input is adequately validated and sanitized
* [x] all publicly displayed strings are internationalized (usually using `esc_html__()`, see https://codex.wordpress.org/I18n_for_WordPress_Developers)
* [x] My code is tested.
* [x] My code follows the Event Espresso code style.
* [x] My code has proper inline documentation.
* [x] My code accounts for when the site is in Maintenance Mode (MM2 especially disallows usage of models)

"