Bumping version to 4.9.80.rc.008. Previous Commit message: "Add reset state actions to our stores. (#998)
## Problem this Pull Request solves

While working on the demo app (#988) and the Barcode Scanner app (https://github.com/eventespresso/eea-barcode-scanner/pull/7) I realized that there will be need for some sort of reset state action for our stores that allows dispatches to restore state to its default shape.  For instance with the barcode scanner app the user might scan a new registration for checkin that we'll want to reset all the tracked related registrations etc in the state.

Note: a [pull I did for GB](https://github.com/WordPress/gutenberg/pull/14225) recently that adds more sledgehammer like control over the resolution cache in the `core/data` store.  It will make it easier for resetting that state in this pull.  The downside is it will be tied to a version of WordPress so I can only use it IF the functions exist.
"