rerender ticket item when calculator modal is closed.
This is necessary because the calculator mutates the ticket.total but the grid item for the ticket is never rerendered so the updated total for the ticket entity is not reflected in the dom (state doesn’t change when the ticket entity is mutated directly).

It’s actually GOOD that we mutate the ticket entity because if state changed everytime a field was changed the ticket item would be re-rendering a LOT.  With the technique implemented here, we only do a rerender of the grid item when the calculator modal is closed. (note, in a later commit this will need done for list item too).