pass along refreshed prop for triggering update on ticket details
This is needed because ticket details might get updated elsewhere in the react tree (edit ticket modal as an example) but the components in the details are wrapped in `useMemo` and linked to `ticket` as a dependency.  Since the ticket entity is mutated and not changed, the memo’d components wouldn’t rerender.  Passing along `refreshed` when ticket details are updated ensures the ticket details components rerender.