fix tests
the `@wordpress/data` store has some changes where `core/data` is now a virtual store.  That means that in the test environment `select( ‘core/data’ )` returns nothing.  So for tests against code using that store (for resolution state), we need to initialize the virtual store for mocks used in the tests.

Note: eventually this may be something we have to address in our code once the virtual store is deprecated.

js_only