Fix CPT types to accept multiple sources for field resolvers
CPTs can be resolved by WP GraphQL which passes its own custom instance of WP_Post. They can also be resolved by the custom connections between custom types or other CPTs, which pass a model instance of the CPT type.

Also, the parent for CPTs is already taken care of by WP GraphQL, so we don't need to resolve it.