Bumping version to 4.10.8.rc.002. Previous Commit message: "Allow multiple categories to be used in ESPRESSO_EVENTS (#3125)
* Switch to sanitize_text_field() so that dashes aren't stripped from the categories passed.

* Build out $event_category_slugs_array from whatever is passed and trim each element in the array to remove whitespaces

Use $event_category_slugs_array to build out an placeholder string to be used with prepare. (Count the number of slugs and add N number of placeholders to a string)

Switch to use IN in the SQL query and pass both $event_category_slugs_prepare and $event_category_slugs_array to prepare.

* Add braces around $event_category_slugs_prepare

* Style fixes."